\name{CardinaliView-package}
\alias{CardinaliView-package}
\alias{CardinaliView}
\docType{package}

\title{Graphical user interface for the Cardinal mass spectrometry imaging package}

\description{
A graphical user interface for Cardinal: visualization and exploration of mass spectrometry imaging experiments.
}

\details{
CardinaliView is a GUI for the Cardinal package for mass spectrometry imaging. Its current implementation only supports exploration and visualization of ion images and mass spectra, but support for the rest of the pre-processing and statistical analysis tools in Cardinal are in development.

To get started, type \code{CardinaliView()}.
}

\author{
Kyle D. Bemis

Maintainer: Kyle D. Bemis <kbemis@purdue.edu>
}
\seealso{
	\code{\link[CardinaliView]{iView}}
}
\keyword{package}
